BugFix: type of create-event needs to be integer between 1 and 3.

+++ src/routes/create-event.ts
@@ -11,7 +11,7 @@ export async function createEvent(app: FastifyInstance) {
       schema: {
         body: z.object({
           name: z.string(),
-          type: z.number().min(1).max(3), // 1 - receipt | 2 - discount | 3 - information
+          type: z.number().int().min(1).max(3), // 1 - receipt | 2 - discount | 3 - information
         }),
       },
       onRequest: [verifyJwt],
